backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Python backend for dodgechat"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "app", from = "." }]

[tool.poetry.dependencies]
alembic = "==1.7.7"
black = "==22.3.0"
bleach = "==5.0.0"
blinker = "==1.4"
click = "==8.1.3"
coverage = "==6.3.2"
dnspython = "==2.2.1"
dominate = "==2.6.0"
email-validator = "==1.2.1"
idna = "==3.3"
importlib-metadata = "==6.8.0"
itsdangerous = "==2.0.1"
marshmallow = "==3.14.1"
mypy-extensions = "==0.4.3"
pathspec = "==0.9.0"
platformdirs = "==2.5.2"
psycopg2 = "==2.9.3"
pyparsing = "==3.0.8"
python = "^3.9"
python-dotenv = "==0.20.0"
six = "==1.16.0"
tomli = "==2.0.1"
typing-extensions = "==4.2.0"
visitor = "==0.1.3"
webencodings = "==0.5.1"
zipp = "==3.8.0"
Flask = "==2.1.2"
Flask-Bootstrap = "==3.3.7.1"
Flask-HTTPAuth = "==4.6.0"
Flask-Login = "==0.6.1"
Flask-Mail = "==0.9.1"
Flask-Migrate = "==3.1.0"
Flask-Moment = "==1.0.2"
Flask-PageDown = "==0.4.0"
Flask-SQLAlchemy = "==2.5.1"
Flask-SSLify = "==0.1.5"
Flask-WTF = "==1.0.1"
Jinja2 = "==3.1.2"
Mako = "==1.2.0"
Markdown = "==3.3.7"
MarkupSafe = "==2.1.1"
SQLAlchemy = "==1.4.36"
Werkzeug = "==2.1.2"
WTForms = "==3.0.1"
pytest = "==7.1.2"
faker = "==13.11.1"
flask-cors = "==3.0.10"
flask-socketio = "==5.2.0"
gunicorn = "==21.2.0"
eventlet = "==0.33.2"
redis = "==4.4.0"
flask-session = "==0.4.0"
celery = "==5.3.4"

[tool.poetry.dev-dependencies]
black = "*"

[build-system]
# Should be the same as `$POETRY_VERSION`:
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"
